+++ src/App.jsx
@@ -1,22 +1,13 @@
 import { RouterProvider } from 'react-router-dom';
 import './App.css';
-import { Suspense, useEffect } from 'react';
+import { Suspense } from 'react';
 import router from '@/routes';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import useUserStore from '@zustand/store';
 
 // react-query 사용
 const queryClient = new QueryClient();
 
 function App() {
-  const { setProducts, setCartProducts } = useUserStore();
-
-  useEffect(() => {
-    console.log('APP JS');
-    setProducts();
-    setCartProducts();
-  }, []);
-
   return (
     <>
       <QueryClientProvider client={queryClient}>
